move file commands into subcommand
