+++ Sources/todo-file/ReadCommand.swift
@@ -1,0 +1,16 @@
+import Commandant
+import Result
+
+struct ReadCommand: CommandProtocol {
+  typealias Options = NoOptions<String>
+
+  let verb = "read"
+  let function = "Read task storage file"
+
+  func run(_ options: Options) -> Result<Void, String> {
+    return Result {
+      try jsonStorage.load().print()
+    }
+  }
+}
+
