+++ Sources/todo/main.swift
@@ -1,6 +1,5 @@
 import Commandant
 import Foundation
-import TaskManagement
 
 let commands = CommandRegistry<String>()
 
@@ -9,7 +8,6 @@ commands.register(AddCommand())
 commands.register(RemoveCommand())
 commands.register(DoCommand())
 commands.register(EditCommand())
-commands.register(FileCommand())
 commands.register(HelpCommand(registry: commands))
 
 commands.main(defaultVerb: ListCommand().verb) { error in
